requirements.txt
Django==2.2.1
pytz==2019.1
httplib2==0.12.3
feedparser==5.2.1
Markdown==2.6.11 # pyup: < 3
psycopg2==2.8.2
Pillow==6.0.0
versiontools==1.9.1
pep8==1.7.1
pyflakes==2.1.1
mccabe==0.6.1
configparser==3.7.4
pycodestyle==2.5.0
flake8==3.7.7
tornado==6.0.2
beautifulsoup4==4.7.1
cssselect==1.0.3
lxml==4.3.3
fuzzywuzzy==0.17.0
sure==1.4.11
traitlets==4.3.2  # ipython
ipython_genutils==0.2.0  # ipython
decorator==4.4.0  # ipython
ptyprocess==0.6.0  # ipython
pexpect==4.7.0  # ipython
pathlib2==2.3.3  # ipython
pickleshare==0.7.5  # ipython
simplegeneric==0.8.1  # ipython
path.py==12.0.1  # ipython
backports.shutil_get_terminal_size==1.0.0  # ipython
wcwidth==0.1.7  # ipython
prompt_toolkit==2.0.9  # ipython
pygments==2.3.1  # ipython
scandir==1.10.0  # ipython
ipython==7.5.0
ipdb==0.12
coverage==4.5.3
pyasn1==0.4.5
cryptography==2.6.1  # pyOpenSSL
pyOpenSSL==19.0.0
ndg-httpsclient==0.5.1
urllib3==1.25.2  # requests
chardet==3.0.4  # requests
idna==2.8  # requests
certifi==2019.3.9  # requests
requests==2.21.0
logilab-common==1.4.2
logilab-astng==0.24.3
astroid==2.2.5
pylint==2.3.1
six==1.12.0
ipaddress==1.0.22  # faker
faker==1.0.5  # factory_boy
factory_boy==2.11.1
ldap3==2.6
sqlparse==0.3.0
python-dateutil==2.8.0

django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-annoying==0.10.4
django-appconf==1.0.3
rcssmin==1.0.6  # django-compressor
rjsmin==1.0.12  # django-compressor
django-compressor==2.2
django-bootstrap4==0.0.8
django-debug-toolbar==1.11
django-waffle==0.16.0
django-jenkins==0.110.0
typing==3.6.6  # For django-extensions
django-extensions==2.1.6
django-impersonate==1.4
django-markwhat==1.6.1
django-flatblocks==0.9.4

djangorestframework==3.9.3

text_unidecode==1.2

django-taggit==1.1.0

nameparser==1.0.3
oauth2==1.9.0.post1
oauthlib==3.0.1

pyparsing==2.4.0
edtf==3.0.0 # pyup: <4.0.0

pbr==5.2.0 # bandit
pyyaml==5.1 # bandit
stevedore==1.30.1  # bandit
bandit==1.5.1

entrypoints==0.3
